Register file clocks in data on falling edge.
This fixes a race condition when a instruction writes back data to the
register file at the same cycle a instruction in id needs that register.
If we're unlucky the updated value will not be able to propagate the
register file, into the pipeline register during the rising edge event.

We have to make sure the wb stage does not grow so that we can safely
assume that the data is stable by the falling edge (half a cycle).